Updated provision script
- Use different php.ini. This has a post and upload max filesize of 1024M
- Copy .my.cnf to /home/vagrant
- Updated Vagrantfile provision script
- Added files to gitignore to ignore stuff

+++ Vagrantfile
@@ -2,11 +2,29 @@
 # vi: set ft=ruby :
 
 $installscript = <<SCRIPT
-echo Copying default site config /var/www/000-default.conf to /etc/apache2/sites-available/000-default.conf
-cp /var/www/000-default.conf /etc/apache2/sites-available/
 
-echo Making new www directory in /var/www/public/www
-mkdir /var/www/public/www
+if [ ! -f "/var/www/.provisioned_defaultvhost" ]; then
+    echo Copying default site config /var/www/000-default.conf to /etc/apache2/sites-available/000-default.conf
+    cp /var/www/000-default.conf /etc/apache2/sites-available/
+    touch /var/www/.provisioned_defaultvhost
+fi
+
+if [ ! -f "/var/www/.provisioned_phpini" ]; then
+    echo Copying php.ini file with 1024M upload_max_filesize and post_max_size.
+    cp /var/www/php.ini /etc/php/7.0/apache2/
+    touch /var/www/.provisioned_phpini
+fi
+
+if [ ! -f "/var/www/.provisioned_mycnf" ]; then
+    echo Copying .my.cnf file
+    cp /var/www/.my.cnf /home/vagrant/
+    touch /var/www/.provisioned_mycnf
+fi
+
+if [ ! -d "/var/www/public/www" ]; then
+    echo Making new www directory in /var/www/public/www
+    mkdir /var/www/public/www
+fi
 
 echo Reloading Apache2
 service apache2 reload
